refactor: 패키지 이름 및 구조 수정
