+++ app/src/main/java/nextstep/jwp/controller/AbstractController.java
@@ -1,7 +1,7 @@
 package nextstep.jwp.controller;
 
-import nextstep.jwp.model.HttpRequest;
-import nextstep.jwp.model.HttpResponse;
+import nextstep.jwp.domain.request.HttpRequest;
+import nextstep.jwp.domain.response.HttpResponse;
 
 import java.io.IOException;
