+++ app/src/main/java/nextstep/jwp/controller/DefaultController.java
@@ -1,8 +1,8 @@
 package nextstep.jwp.controller;
 
-import nextstep.jwp.model.HttpRequest;
-import nextstep.jwp.model.HttpResponse;
-import nextstep.jwp.model.HttpStatus;
+import nextstep.jwp.domain.request.HttpRequest;
+import nextstep.jwp.domain.response.HttpResponse;
+import nextstep.jwp.domain.response.HttpStatus;
 
 import java.util.LinkedHashMap;
 import java.util.Map;
