+++ app/src/main/java/nextstep/jwp/domain/ContentType.java
@@ -1,4 +1,4 @@
-package nextstep.jwp.model;
+package nextstep.jwp.domain;
 
 import java.util.Arrays;
