+++ app/src/main/java/nextstep/jwp/domain/User.java
@@ -1,4 +1,4 @@
-package nextstep.jwp.model;
+package nextstep.jwp.domain;
 
 public class User {
